move camera rendering to dedicated class
